Super Flappy Bird
My Minimum Viable Project will be the Flappy Bird Game where the bird representing the user has to dodge the randomly generated pipes on the screen. The score of the user keeps increasing until the bird hits an obstacle. Once an obstacle is encountered, the game resets itself and places the bird back at the center of the screen with a score of 0. The user has three lives before the Game Over Screen is displayed.
To go beyond the MVP, my game will allow users to choose the universe they want to play in and each universe will have a different implementation of the flappy bird game.